+++ AccountBook/Utils/Extension.swift
@@ -29,49 +29,6 @@ extension Int {
     }
 }
 
-extension String {
-    func convert(_ value: String) -> SpendType.RawValue {
-        switch value {
-        case "대중교통":
-            return SpendType.대중교통.rawValue
-        case "식사":
-            return SpendType.식사.rawValue
-        case "보험":
-            return SpendType.보험.rawValue
-        case "술자리":
-            return SpendType.술자리.rawValue
-        case "커피":
-            return SpendType.커피.rawValue
-        case "물건구입":
-            return SpendType.물건구입.rawValue
-        default:
-            return SpendType.기타.rawValue
-        }
-    }
-    
-    func toDate(_ time: String) -> Date {
-        let dateString:String = time
-
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        dateFormatter.timeZone = NSTimeZone(name: "UTC") as TimeZone?
-
-        return dateFormatter.date(from: dateString)!
-    }
-    
-    func image() -> UIImage? {
-        let size = CGSize(width: 35, height: 40)
-        UIGraphicsBeginImageContextWithOptions(size, false, 0)
-        UIColor.white.set()
-        let rect = CGRect(origin: .zero, size: size)
-        UIRectFill(CGRect(origin: .zero, size: size))
-        (self as AnyObject).draw(in: rect, withAttributes: [.font: UIFont.systemFont(ofSize: 30)])
-        let image = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return image
-    }
-}
-
 extension UIView {
     @IBInspectable var cornerRadius: CGFloat {
         get {
